+++ components/search/common/index.ts
@@ -1,0 +1,4 @@
+export { SearchPage } from "./SearchPage"
+export type { SortOptionInput, SearchType } from "./SearchPage"
+export type { FilterPanelConfig } from "./FilterPanel"
+export { getServerConfig, server } from "./config"
