+++ public/locales/en/search.json
@@ -23,7 +23,8 @@
     "hearing": {
       "month": "Month",
       "year": "Year",
-      "chairNames": "Chair"
+      "chairNames": "Chair",
+      "committeeName": "Committee"
     }
   },
   "sort_by": {
@@ -42,12 +43,13 @@
     "placeholder": {
       "bills": "Search For Bills",
       "testimony": "Search For Testimony",
-      "hearings": "Search For Hearings"
+      "hearing": "Search For Hearings"
     }
   },
   "search_error": "Something went wrong. Please try again. Original message: {{error}}",
   "video_available": "Video available",
   "location_label": "Location",
   "agenda_label": "Agenda Topics",
-  "bills_label": "Bills"
+  "bills_label": "Bills",
+  "more_bills": "and {{count}} others"
 }
